Incorporate prime finding function
