+++ client/index.html
@@ -11,6 +11,10 @@
   HELLO I AM A WORKER BEE
   </div>
 
+  <div>
+    Result so far: <span id="result">In progress...</span>
+  </div>
+
 </head>
 <body>
