+++ client/index.js
@@ -1,5 +1,30 @@
 $(document).ready(function() {
 
+  /************************************************
+  // Function that finds primes between two values
+  ************************************************/
+  var findPrimes = function(min, max) {
+    var primeTester = function(n) {
+      for (var i = 2; i < n - 1; i++) {
+        if (n % i === 0) {
+          return false;
+        }
+      }
+      return true;
+    };
+
+    var result = [];
+    for (var i = min; i <= max; i++) {
+      if (primeTester(i)) {
+        result.push(i);
+      }
+    }
+    return result;
+  }
+
+  /************************************************
+  // Socketing
+  ************************************************/
   var socket = io();
 
   var sendReady = function() {
@@ -9,15 +34,17 @@ $(document).ready(function() {
   socket.on('newjob', function(job) {
     console.log('Working on new job');
     
-    var functionSnippet = eval('(' + job.func + ')');
-    var result = functionSnippet(job.data[0], job.data[1]);
+    var result = findPrimes(job.data[0], job.data[1]);
     
     console.log('Job complete. Result is: ', result);
     console.log('Sending result back to server');
     socket.emit('jobdone', result);
-
   });
 
-  setTimeout(sendReady, 2000);
+  socket.on('jobresult', function(result) {
+    $('#result').text(result);
+  })
+
+  sendReady();
 
 });
