+++ package.json
@@ -20,6 +20,8 @@
     "body-parser": "^1.15.2",
     "express": "^4.14.0",
     "http": "0.0.0",
+    "lodash": "^4.14.1",
+    "node-timers": "0.0.1",
     "socket.io": "^1.4.8"
   }
 }
